:sparkles: Fix integration tests
