+++ src/main/kotlin/com/unqueam/gamingplatform/infrastructure/configuration/EmailBeans.kt
@@ -1,0 +1,31 @@
+package com.unqueam.gamingplatform.infrastructure.configuration
+
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Profile
+import org.springframework.mail.javamail.JavaMailSender
+import org.springframework.mail.javamail.JavaMailSenderImpl
+
+@Configuration
+@Profile("prod")
+class EmailBeans {
+
+    @Value("\${EMAIL_PASSWORD}")
+    private lateinit var emailPassword: String
+
+    @Bean
+    fun javaMailSender(): JavaMailSender {
+        val mailSender = JavaMailSenderImpl()
+        mailSender.host = "smtp.gmail.com"
+        mailSender.port = 587
+        mailSender.username = "unqueam.platform@gmail.com"
+        mailSender.password = emailPassword
+        val props = mailSender.javaMailProperties
+        props["mail.transport.protocol"] = "smtp"
+        props["mail.smtp.auth"] = "true"
+        props["mail.smtp.starttls.enable"] = "true"
+        props["mail.debug"] = "true"
+        return mailSender
+    }
+}
